+++ pkg/diag/colors/colors.go
@@ -145,9 +145,10 @@ var (
 	SpecError   = Red             // for errors.
 	SpecWarning = Yellow          // for warnings.
 
-	SpecHeadline  = BrightMagenta + Bold // for headings in the CLI.
-	SpecPrompt    = Cyan + Bold          // for prompting the user
-	SpecAttention = BrightRed            // for messages that are meant to grab attention.
+	SpecHeadline    = BrightMagenta + Bold // for headings in the CLI.
+	SpecSubHeadline = Bold                 // for subheadings in the CLI.
+	SpecPrompt      = Cyan + Bold          // for prompting the user.
+	SpecAttention   = BrightRed            // for messages that are meant to grab attention.
 
 	// for simple notes.  Just use the standard terminal text color.
 	SpecNote = Reset